Refactor wallet service to use integer balances
Changed wallet and transaction models to use INTEGER instead of DECIMAL for balance-related fields to simplify calculations and avoid floating point issues. Updated controller and repository logic to ensure amounts are handled as numbers. Also fixed import paths and removed unused contest event handler references. Added wallet_db_data volume to docker-compose for wallet service persistence.

+++ apps/wallet_service/src/models/transaction.model.ts
@@ -76,17 +76,17 @@ export default function (sequelize: Sequelize) {
       },
 
       amount: {
-        type: DataTypes.DECIMAL(10, 2),
+        type: DataTypes.INTEGER,
         allowNull: false,
       },
 
       balanceBefore: {
-        type: DataTypes.DECIMAL(10, 2),
+        type: DataTypes.INTEGER,
         allowNull: false,
       },
 
       balanceAfter: {
-        type: DataTypes.DECIMAL(10, 2),
+        type: DataTypes.INTEGER,
         allowNull: false,
       },
